Prevent duplicated package versions
The function get_versions_for_package_from_repo might return
duplicated package versions in case the specified version is present
both as wheel and as sdist package.
Duplicated entries cause the package resolution to take longer.
The new behavior is that the package version is considered if the
package is present as either whel or sdist package.

Signed-off-by: Stefano Bennati <stefano.bennati@here.com>