Fix `is_valid_version()` for prerelease versions
By default, the check if an empty specifier contains a prerelease
version returns `false`. This made the `is_valid_version()` check for
prerelease versions fail, when there is also a requirement with a
non-empty specifier. In this case, the previous logic detected a
conflict of the prerelease version and the empty specifier.

Fix this by explicitly allowing prerelease versions when checking if a
specifier contains a version.

Note that the function that collects candidates for a dependency still
ignores prerelease versions if any release versions are available, as
required by the specs [1].

[1]: https://packaging.python.org/en/latest/specifications/version-specifiers/#handling-of-pre-releases

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>