Fix resolution issues (#248)
* Update expected test results

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>

* Test all linux platforms defined in utils_pypi

Use the list of linux platforms as defined in utils_pypi for testing,
instead of redefining the list in the test. This ensures that newly
added platforms are tested and fixes the problem that the list in the
test was already outdated.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>

* Add manylinux_2_31_x86_64 platform

For example, the platform is used by open3d 19.0.0 [1].

[1]: https://pypi.org/project/open3d/0.19.0/#files

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>

* Fix `is_valid_version()` for prerelease versions

By default, the check if an empty specifier contains a prerelease
version returns `false`. This made the `is_valid_version()` check for
prerelease versions fail, when there is also a requirement with a
non-empty specifier. In this case, the previous logic detected a
conflict of the prerelease version and the empty specifier.

Fix this by explicitly allowing prerelease versions when checking if a
specifier contains a version.

Note that the function that collects candidates for a dependency still
ignores prerelease versions if any release versions are available, as
required by the specs [1].

[1]: https://packaging.python.org/en/latest/specifications/version-specifiers/#handling-of-pre-releases

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>

---------

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>