Filter out empty values from `install_requires`
Sometimes, when the `--analyze-setup-py-insecurely` parameter is used,
the requirements resolution can contain some empty elements. For
instance for gnupg-2.3.1/gnupg-2.3.1/setup.py, the requirements are
resolved to ['psutil>=1.2.1', '']. The second element of this array
crashes the Python Inspector.
This commit fixes the issue by removing empty elements from the array.

Resolves #247.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.com>