Ensure that the cached file is not empty before using it
We've observed multiple cases where a zero-byte file is cached, leading
to crashes with the error message:

    zipfile.BadZipFile: File is not a zip file

I was not able to reproduce this issue reliably enough to identify the
root cause. However, adding a check to ensure that the cached file is
not empty before using it will mitigate the problem.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.com>